Refatored code and fix bugs

+++ WebDAVServer/WebDAVServer/src/file/FileManager.cs
@@ -9,6 +9,9 @@ namespace WebDAVServer.file {
         private readonly String mRoot;
 
         internal static void init(String root) {
+            if (null == root) {
+                throw new ArgumentNullException("root");
+            }
             instance = new FileManager(root);
         }
 
@@ -20,50 +23,95 @@ namespace WebDAVServer.file {
         }
 
         private FileManager(String root) {
+            if (null == root) {
+                throw new ArgumentNullException("root");
+            }
             mRoot = root;
         }
 
         internal FileStream getFile(String url) {
+            if (null == url) {
+                throw new ArgumentNullException("url");
+            }
             var path = mRoot + url;
             return File.Open(path, FileMode.Open);
         }
 
         internal FileStream getFileForRead(String url) {
+            if (null == url) {
+                throw new ArgumentNullException("url");
+            }
             var path = mRoot + url;
             return File.OpenRead(path);
         }
 
         internal FileInfo getFileInfo(String url) {
+            if (null == url) {
+                throw new ArgumentNullException("url");
+            }
             var path = mRoot + url;
             return new FileInfo(path);
         }
 
         internal DirectoryInfo getDirInfo(String url) {
+            if (null == url) {
+                throw new ArgumentNullException("url");
+            }
             var path = mRoot + url;
             return new DirectoryInfo(path);
         }
 
         internal FileStream createFile(String url) {
+            if (null == url) {
+                throw new ArgumentNullException("url");
+            }
             var path = mRoot + url;
             return File.Open(path, FileMode.Create);
         }
 
         internal bool copyFile(String src, String dst, bool overwrite) {
+            if (null == src) {
+                throw new ArgumentNullException("src");
+            }
+            if (null == dst) {
+                throw new ArgumentNullException("dst");
+            }
             var res = !getFileInfo(dst).Exists;
             File.Copy(mRoot + src, mRoot + dst, overwrite);
             return res;
         }
 
         internal bool moveFile(String src, String dst, bool overwrite) {
-            var res = !getFileInfo(dst).Exists;
-            if (overwrite) {
-                File.Delete(mRoot + dst);
+            if (null == src) {
+                throw new ArgumentNullException("src");
             }
-            File.Move(mRoot + src, mRoot + dst);
-            return res;
+            if (null == dst) {
+                throw new ArgumentNullException("dst");
+            }
+            if (getFileInfo(src).Exists) {
+
+                var res = !getFileInfo(dst).Exists;
+                if (overwrite) {
+                    File.Delete(mRoot + dst);
+                }
+                File.Move(mRoot + src, mRoot + dst);
+                return res;
+            }
+            if (!getDirInfo(src).Exists) {
+                throw new FileNotFoundException(src);
+            }
+            var dirRes = !getDirInfo(dst).Exists;
+            if (overwrite) {
+                Directory.Delete(mRoot + dst);
+            }
+            Directory.Move(mRoot + src, mRoot + dst);
+            return dirRes;
         }
 
         internal void deleteFileOrDir(String url) {
+            if (null == url) {
+                throw new ArgumentNullException("url");
+            }
             var path = mRoot + url;
             var file = new FileInfo(path);
             if (file.Exists) {
